+++ ws/utils.py
@@ -1,0 +1,14 @@
+import json
+
+from channels.generic.websocket import AsyncJsonWebsocketConsumer
+
+
+class MyJsonWebsocketConsumer(AsyncJsonWebsocketConsumer):
+    async def send_json(self, content, close=False):
+        """
+        unicode 编码 json 并发给客户端
+        """
+        await super().send(text_data=json.dumps(content, ensure_ascii=False), close=close)
+
+    async def notification(self, event):
+        await self.send_json(event['content'])
